Some cleanup varaible name, sample file
