+++ .env_sample
@@ -1,2 +1,21 @@
-#PORT
-PORT = 443
+# APP PORT
+APP_PORT = 443
+
+# SESSION SECRET
+SESSION_SECRET = YOUR_SESSION_SECRET
+
+# This could surely be better managed
+# DB CREDENTIALS FOR PRODUCTION
+DB_USERNAME = PROD_DB_USERNAME
+DB_PASSWORD = PROD_DB_PASSWORD
+DB_NAME = PROD_DB_PASSWORD 
+DB_HOST = PROD_DB_CONTAINER_NAME
+
+# DB CREDENTIALS FOR DEVELOPMENT *** Probably a good idea to completely remove in prod
+DEV_DB_USERNAME = app_manager
+DEV_DB_PASSWORD = M^n2AqDqyUMu=FB+
+DEV_DB_NAME = caption_contest_test
+DEV_DB_HOST = postgres-test
+
+# BCRYPT Number of SALT ROUNDS
+SALT_ROUND = 12
