+++ database/config/config.js
@@ -1,16 +1,16 @@
 module.exports = {
   development: {
-    username: process.env.DEV_DATABASE_USERNAME,
-    password: process.env.DEV_DATABASE_PASSWORD,
-    database: process.env.DEV_DATABASE_NAME,
-    host: process.env.DEV_DATABASE_HOST,
+    username: process.env.DEV_DB_USERNAME,
+    password: process.env.DEV_DB_PASSWORD,
+    database: process.env.DEV_DB_NAME,
+    host: process.env.DEV_DB_HOST,
     dialect: 'postgres'
   },
   production: {
-    username: process.env.DATABASE_USERNAME,
-    password: process.env.DATABASE_PASSWORD,
-    database: process.env.DATABASE_NAME,
-    host: process.env.DATABASE_HOST,
+    username: process.env.DB_USERNAME,
+    password: process.env.DB_PASSWORD,
+    database: process.env.DB_NAME,
+    host: process.env.DB_HOST,
     dialect: 'postgres'
   }
 }
