+++ middleware/authentication.js
@@ -1,0 +1,9 @@
+const isAuthenticated = (req, res, next) => {
+    if ( req.isAuthenticated() ) {
+      next()
+    } else {
+      res.redirect('/login')
+    }
+}
+
+module.exports = isAuthenticated
